# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Our Voice Our Rights API
  version: 1.0.0
  description: API for district-level MGNREGA metrics and geo detection.
servers:
  - url: http://localhost:8080
paths:
  /api/states:
    get:
      summary: List states
      operationId: listStates
      tags: [States]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateDto'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/districts:
    get:
      summary: List districts in a state
      operationId: listDistricts
      tags: [Districts]
      parameters:
        - in: query
          name: state
          required: true
          schema:
            type: string
          description: State name or code
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistrictDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/performance/{district}:
    get:
      summary: Latest performance for a district
      operationId: latestPerformance
      tags: [Performance]
      parameters:
        - in: path
          name: district
          required: true
          schema:
            type: string
        - in: query
          name: fin_year
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/performance/{district}/history:
    get:
      summary: Historical performance for a district
      operationId: performanceHistory
      tags: [Performance]
      parameters:
        - in: path
          name: district
          required: true
          schema:
            type: string
        - in: query
          name: fin_year
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerformanceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/compare:
    get:
      summary: Compare two districts
      operationId: compareDistricts
      tags: [Comparison]
      parameters:
        - in: query
          name: district1
          required: true
          schema:
            type: string
        - in: query
          name: district2
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/geo/detect:
    get:
      summary: Detect district from coordinates or IP
      operationId: detectLocation
      tags: [Geo]
      parameters:
        - in: query
          name: lat
          required: false
          schema:
            type: number
            format: double
        - in: query
          name: lon
          required: false
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoDetectDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/geo/reverse:
    get:
      summary: Reverse geocode coordinates
      operationId: reverseGeocode
      tags: [Geo]
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: lon
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoReverseDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/health/ingestion:
    get:
      summary: Health status of ingestion pipeline
      operationId: ingestionHealth
      tags: [Health]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  schemas:
    StateDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
    DistrictDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        stateId:
          type: integer
          format: int64
    PerformanceDto:
      type: object
      properties:
        district:
          type: string
        state:
          type: string
        finYear:
          type: string
        month:
          type: integer
        totalPersondays:
          type: integer
          format: int64
        totalHouseholds:
          type: integer
          format: int64
        expenditure:
          type: number
          format: double
        updatedAt:
          type: string
          format: date-time
    CompareDto:
      type: object
      properties:
        district1:
          $ref: '#/components/schemas/PerformanceDto'
        district2:
          $ref: '#/components/schemas/PerformanceDto'
        totalPersondaysDelta:
          type: integer
          format: int64
        totalHouseholdsDelta:
          type: integer
          format: int64
        expenditureDelta:
          type: number
          format: double
    GeoDetectDto:
      type: object
      properties:
        state:
          type: string
        district:
          type: string
        method:
          type: string
        accuracyRadiusKm:
          type: number
          format: double
    GeoReverseDto:
      type: object
      properties:
        address:
          type: string
        state:
          type: string
        district:
          type: string
        method:
          type: string
        accuracyRadiusKm:
          type: number
          format: double
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    ServiceUnavailable:
      description: Upstream unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'